services/ticket-service/requirements.txt
fastapi>=0.95.1
uvicorn>=0.22.0
pymongo>=4.3.3
pydantic>=2.0.0
python-jose>=3.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0